Extract DOM selectors and split side effects
All of the side effects (except for a few to be captured in cut/paste) are
now wrapped in streams, handled either by emitting actions in `events`
or updating the DOM in `render`.
