+++ client/editor/instance/index.hbs
@@ -5,7 +5,6 @@
             data-brk-onfocus="onFocus"
             data-brk-oninput="onInput"
             data-brk-onkeydown="onKeydown"
-            data-brk-onkeyup="onKeyup"
             data-brk-onkeypress="onKeypress"
             data-brk-oncut="onCut"
             data-brk-onpaste="onPaste"
