+++ client/editor/instance/index.js
@@ -7,21 +7,14 @@ import Prism from '../../prism';
 import component from 'brookjs/component';
 import render from 'brookjs/render';
 import events from 'brookjs/events';
+import { selectSelectionStart, selectSelectionEnd } from '../../selector';
 import template from './index.hbs';
 
-const CRLF = /\r?\n|\r/g;
+// const CRLF = /\r?\n|\r/g;
 
-const keyupIgnore = [
-    9, 91, 93, 16, 17, 18, // modifiers
-    20, // caps lock
-    13, // Enter (handled by keydown)
-    112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, // F[0-12]
-    27 // Esc
-];
-
-const updateLinenumber = pre => stream(emitter => {
-    let content = pre.textContent;
-    let ss = pre.selectionStart;
+const updateLinenumber = (/* pre */) => stream(emitter => {
+    // let content = pre.textContent;
+    // let ss = pre.selectionStart;
     // let se = pre.selectionEnd;
     //
     // // @todo push into store
@@ -29,9 +22,9 @@ const updateLinenumber = pre => stream(emitter => {
     // se && pre.setAttribute('data-se', se);
 
     // Update current line highlight
-    let line = (content.slice(0, ss).match(CRLF) || []).length;
+    // let line = (content.slice(0, ss).match(CRLF) || []).length;
 
-    pre.setAttribute('data-line', line + 1);
+    // pre.setAttribute('data-line', line + 1);
 
     emitter.end();
 });
@@ -39,7 +32,9 @@ const updateLinenumber = pre => stream(emitter => {
 const mapKeydownToAction = evt => {
     const { altKey, shiftKey: inverse, metaKey, ctrlKey } = evt;
     const cmdOrCtrl = metaKey || ctrlKey;
-    const { selectionStart: ss, selectionEnd: se, textContent: value } = evt.target;
+    const { textContent: value } = evt.target;
+    const ss = selectSelectionStart(evt.target);
+    const se = selectSelectionEnd(evt.target);
     const length = ss === se ? 1 : Math.abs(se - ss);
     const start = se - length;
     const cursor = [ss, se];
@@ -91,6 +86,102 @@ const mapCutEventToAction = evt => {
 
 const renderTemplate = render(template);
 
+const setSelectionRange = (node, ss, se) => stream(emitter => {
+    const loop = requestAnimationFrame(() => {
+        let range = document.createRange();
+        let offset = findOffset(node, ss);
+
+        range.setStart(offset.element, offset.offset);
+
+        if (se && se !== ss) {
+            offset = findOffset(node, se);
+        }
+
+        range.setEnd(offset.element, offset.offset);
+
+        let selection = window.getSelection();
+        selection.removeAllRanges();
+        selection.addRange(range);
+
+        emitter.end();
+    });
+
+    return () => cancelAnimationFrame(loop);
+});
+
+function findOffset(root, ss) {
+    let container;
+
+    if (!root) {
+        return null;
+    }
+
+    let offset = 0;
+    let element = root;
+
+    do {
+        container = element;
+        element = element.firstChild;
+
+        if (element) {
+            do {
+                var len = element.textContent.length;
+
+                if (offset <= ss && offset + len > ss) {
+                    break;
+                }
+
+                offset += len;
+            } while (element = element.nextSibling);
+        }
+
+        if (!element) {
+            // It's the container's lastChild
+            break;
+        }
+    } while (element && element.hasChildNodes() && element.nodeType !== 3);
+
+    if (element) {
+        return {
+            element: element,
+            offset: ss - offset
+        };
+    } else if (container) {
+        element = container;
+
+        while (element && element.lastChild) {
+            element = element.lastChild;
+        }
+
+        if (element.nodeType === 3) {
+            return {
+                element: element,
+                offset: element.textContent.length
+            };
+        } else {
+            return {
+                element: element,
+                offset: 0
+            };
+        }
+    }
+
+    return {
+        element: root,
+        offset: 0,
+        error: true
+    };
+}
+
+const highlightElement = el => stream(emitter => {
+    let loop = requestAnimationFrame(() => {
+        Prism.highlightElement(el.querySelector('code'), false);
+        emitter.end();
+    });
+
+    return () => cancelAnimationFrame(loop);
+});
+
 export default component({
     onMount: () => stream(emitter => {
         Prism.setAutoloaderPath(__webpack_public_path__);
@@ -99,18 +190,15 @@ export default component({
     }),
     render: R.curry((el, prev, next) => fromPromise(Promise.all([
         Prism.setTheme(next.editor.theme),
-        Prism.togglePlugin('line-highlight', true),
         Prism.togglePlugin('show-invisibles', next.editor.invisibles === 'on')
     ]))
         .ignoreValues()
-        .concat(merge([renderTemplate(el, prev, next), stream(emitter => {
-            let loop = requestAnimationFrame(() => {
-                Prism.highlightElement(el.querySelector('code'), false);
-                emitter.end();
-            });
-
-            return () => cancelAnimationFrame(loop);
-        })]))
+        .concat(merge([
+            renderTemplate(el, prev, next),
+            highlightElement(el),
+            next.instance.cursor ? setSelectionRange(el.querySelector('code'), ...next.instance.cursor) : never(),
+            next.instance.cursor ? updateLinenumber(el.querySelector('pre')) : never()
+        ]))
     ),
     events: events({
         onClick: event$ => event$.flatMap(R.pipe(
@@ -122,47 +210,13 @@ export default component({
             R.map(mapKeydownToAction),
             R.filter(R.identity)
         ),
-        onKeyup: event$ => event$.flatMap(evt => {
-            let pre = evt.target;
-            let keyCode = evt.keyCode || 0;
-            let code = pre.textContent;
-            let lineNumber$ = never();
-
-            if (keyCode < 9 || keyCode === 13 || keyCode > 32 && keyCode < 41) {
-                lineNumber$ = updateLinenumber(pre);
-            }
-
-            if (keyupIgnore.indexOf(keyCode) > -1) {
-                return lineNumber$;
-            }
-
-            return lineNumber$.merge(stream(emitter => {
-                if (keyCode !== 37 && keyCode !== 39) {
-                    let ss = pre.selectionStart;
-                    let se = pre.selectionEnd;
-
-                    Prism.highlightElement(pre);
-
-                    // Dirty fix to #2
-                    if (!/\n$/.test(code)) {
-                        pre.innerHTML = pre.innerHTML + '\n';
-                    }
-
-                    if (ss !== null || se !== null) {
-                        pre.setSelectionRange(ss, se);
-                    }
-                }
-
-                emitter.end();
-            }));
-        }),
         onKeypress: event$ => event$
             .debounce(0)
             .filter(evt => evt.charCode && !(evt.metaKey || evt.ctrlKey))
             .map(evt => {
                 const pre = evt.target;
-                const start = pre.selectionStart;
-                const end = pre.selectionEnd;
+                const start = selectSelectionStart(pre);
+                const end = selectSelectionEnd(pre);
                 const add = String.fromCharCode(evt.charCode);
                 const value = pre.textContent;
                 const del = start === end ? '' : pre.textContent.slice(start, end);
@@ -176,10 +230,10 @@ export default component({
             R.map(mapCutEventToAction)
         ),
         onPaste: event$ => event$.flatMap(evt => stream(emitter => {
-            let pre = evt.target;
-            let ss = pre.selectionStart;
-            let se = pre.selectionEnd;
-            let selection = ss === se ? '' : pre.textContent.slice(ss, se);
+            const pre = evt.target;
+            let ss = selectSelectionStart(pre);
+            let se = selectSelectionEnd(pre);
+            const selection = ss === se ? '' : pre.textContent.slice(ss, se);
 
             if (evt.clipboardData) {
                 evt.preventDefault();
@@ -192,32 +246,31 @@ export default component({
                     value: pre.textContent,
                     add: pasted,
                     del: selection,
-                    start: ss
+                    cursor: [ss, se]
                 }));
 
                 ss += pasted.length;
 
-                pre.setSelectionRange(ss, ss);
+                setSelectionRange(pre, ss, ss);
 
                 pre.onkeyup();
 
                 emitter.end();
             } else {
                 setTimeout(function() {
-                    let newse = pre.selectionEnd;
-
+                    const newse = selectSelectionEnd(pre);
                     let pasted = pre.textContent.slice(ss, newse);
 
                     emitter.value(editorValueChangeAction({
                         value: pre.textContent,
                         add: pasted,
                         del: selection,
-                        start: ss
+                        cursor: [ss, newse]
                     }));
 
                     ss += pasted.length;
 
-                    pre.setSelectionRange(ss, ss);
+                    setSelectionRange(pre, ss, ss);
 
                     pre.onkeyup();
